Refatoração dos models, seperando por apps o usuario, cliente e empresa, optimizando o codigo do projeto
