+++ clientes/admin.py
@@ -1,0 +1,4 @@
+from django.contrib import admin
+from .models import Cliente
+# Register your models here.
+admin.site.register(Cliente)
